# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NGO HomeSuite API
  version: 0.1.0-beta
  description: |
    Starter OpenAPI contract for internal/beta integrations.
    This contract emphasizes stable operationIds and reusable schemas to support
    downstream tooling (client generation, docs, and contract checks).
servers:
  - url: /
paths:
  /api/docs:
    get:
      operationId: getApiDocsIndex
      summary: API docs landing page
      tags: [Docs]
      responses:
        '200':
          description: HTML docs landing page
          content:
            text/html:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/swagger:
    get:
      operationId: getApiSwaggerUi
      summary: Interactive Swagger UI page
      tags: [Docs]
      responses:
        '200':
          description: Swagger UI HTML page
          content:
            text/html:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/openapi.yaml:
    get:
      operationId: getOpenApiSpec
      summary: OpenAPI specification
      tags: [Docs]
      responses:
        '200':
          description: YAML OpenAPI document
          content:
            application/yaml:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /reports/compliance/evidence:
    get:
      operationId: getComplianceEvidence
      summary: Export compliance evidence packet
      tags: [Compliance]
      parameters:
        - in: query
          name: scope
          schema:
            type: string
            enum: [org]
          required: false
          description: Evidence scope (defaults to org)
        - in: query
          name: download
          schema:
            type: integer
            enum: [0, 1]
          required: false
          description: Set to 1 to force file download
      responses:
        '200':
          description: Compliance evidence payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceEvidence'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v2/membership/members:
    get:
      operationId: getV2MembershipMembers
      summary: List members with filter, search, and pagination
      tags: [MembershipV2]
      responses:
        '200':
          description: Membership member list payload
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /ai/minion/chat:
    post:
      operationId: postMinionChat
      summary: Query HomeSuite Minion
      tags: [AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MinionChatRequest'
      responses:
        '200':
          description: Minion response envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MinionChatResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ComplianceEvidence:
      type: object
      properties:
        schema:
          type: string
        sha256:
          type: string
        security_posture:
          type: object
          additionalProperties: true
        data_inventory:
          type: object
          additionalProperties: true
      required: [schema, sha256]
    MinionActionApprovalItem:
      type: object
      properties:
        tool:
          type: string
        token:
          type: string
      required: [tool]
    MinionChatRequest:
      type: object
      properties:
        prompt:
          type: string
        context:
          type: object
          additionalProperties: true
        allow_actions:
          type: boolean
        approved_actions:
          type: array
          items:
            oneOf:
              - type: string
              - $ref: '#/components/schemas/MinionActionApprovalItem'
      required: [prompt]
    SourceSnippet:
      type: object
      additionalProperties: true
    ToolAction:
      type: object
      additionalProperties: true
...
    MinionChatResponse:
      type: object
      properties:
        response:
          type: string
        sources:
          type: array
          items:
            $ref: '#/components/schemas/SourceSnippet'
        actions:
          type: array
          items:
            $ref: '#/components/schemas/ToolAction'
        redactions:
          type: integer
        mode:
          type: string
      required: [response, sources, actions, mode]
...
  responses:
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServiceUnavailable:
      description: Service unavailable or disabled
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    sessionCookieAuth:
      type: apiKey
      in: cookie
      name: session
security:
  - sessionCookieAuth: []